renamed moviesRouter & moviesModel to movieRouter & movieModel
